+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
 ## In Development (0.8.2)
+### New Segments
+- Added **isTrue** segment that can function as both a filter or and evaluation of a field.
+
 ### Improvements
 - Added a check and thow an exception in simplevectordb if the user has clustered and then tries to use cosine for search
 - Added a plugin system so it is easer for external whl files to add commands to chatterlang.